requirements-gpu.txt
tensorflow-gpu
scipy
tqdm
